requirements.txt
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.4.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.4
django-debug-toolbar==3.2.1
django-extensions==3.1.3
django-extra-fields==3.0.2
django-seed==0.2.2
djangorestframework==3.12.4
drf-yasg==1.20.0
entrypoints==0.3
Faker==8.8.2
idna==3.2
inflection==0.5.1
ipykernel==5.5.5
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
packaging==20.9
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.11.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.1.0
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
Send2Trash==1.7.1
six==1.16.0
sqlparse==0.4.1
terminado==0.10.1
testpath==0.5.0
text-unidecode==1.3
tornado==6.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
